The wrapper is no longer a separate process.
Moved the logic in wrap.c into the updater's update() function. This means the
updater can do the update step in a single transaction. Though, I may have
broken stuff if the ldpreloader finds a new node or needs to remove DELETE
flags since it does a create_name_file(). That could be fixed, though.