Send filenames instead of tupid from ldpreload. Also, include the rest of the commit from the wrapper change before. Turns out I suck at git.
Sending the filenames instead of tupids means we don't have to do db accesses
in ldpreload, which avoids the brokenness I introduced last commit. Also, it
avoids all sorts of silly synchronization issues, and actually runs the test
cases about a half second faster, so it is double plus good.