Changed parsing logic for Tupfile dependencies.
Now when a Tupfile tries to use the output from another directory, that other directory will be parsed immediately. This forced the create-node while-loop in the updater to effectively do a while(get a single create node) instead of a get-all-create-nodes (while(!list_empty)), since calling parser_create on a single tupid may cascade to other tupids. This makes both t4004 and t4005 work as intended, in that the Tupfiles can be processed in either initial order (and will be processed in the correct order internally), and will be processed a minimal amount of times.