Moved some of the special directory deletion logic from the monitor to tup_file_mod().
This let's us recursively delete a directory both with and without the monitor. For example, with the monitor not running, you could 'rm -rf foo; tup delete foo' and things would work as expected. I also copied a few of the monitor directory tests into and took out the monitor part, so they verify both cases.