Don't trigger an update if a variable is set to the same value.
Now we check if a variable already exists with the given value when set, so we don't have to do anything in that case. I guess this is somewhat different semantically from what you get when you 'touch' a file, but this makes saving the config much easier for now.