Add t6011 as a bugfix test for premature eval().
It's probably not the best fix, but now a line in a Tupfile is only eval'd if we're not in the false part of an if statement. The issue was the variable settings wouldn't get parsed, but it would still try to use them to eval a line before checking for endif.