Added and fixed t6012.
I guess I don't need to propogate the flags from a node to its children when processing it in the updater, since in a node I only check if the delete flag is set, and if not we can run a command. The delete flag is set properly already when deleting a command (including its output) or a file, so they don't need to propogate. If a command farther down the line is present but doesn't have any flags set, it'll run anyway. Improperly forwarding flags to children nodes caused the issue described in t6012.