Make dependencies on included Tupfiles work.
This is a somewhat largish commit, since I wanted to re-use some of the parsing code for generating name_lists of files to include (in theory you may be able to specify multiple files on an include line, or wildcard them - I have not tried it though). Also I made sure that when you change a Tupfile to stop including another Tupfile (or, stop using a variable) then that dependency is no longer there.