Use a tri-lock mechanism for monitor/program interaction.
I was having issues with the autoupdater, since the monitor would sometimes catch a deletion and think it was outside of the updater part and then try to delete the (already deleted) file itself. The new tri-lock (explained in lock.c) fixes that. Also, I got rid of the updater lock since I don't have a shared lock anymore.