Moved node.flags into its own set of tables for performance.
So when I ran the make_v_tup test up to 100k nodes, I was rather disappointed that the do-nothing update time was 700ms. All tup should be doing is about 3 meaningful selects - on the create nodes, modify nodes, and delete nodes, all of which should be empty. I figured with the indexing that should be pretty quick. When I timed it on the command line, it was almost instantaneous. However, in the code each select takes about 200ms. It turns out the index doesn't help (or is slower, not entirely sure) when you try to match 'flags&?' as opposed to 'flags=?'. Of course, when I want to get all things that have, say, modify flags (1) set, I could do flags=1, flags=3, flags=5, flags=7 etc (anything with the 1 bit set), but that would be dumb. Instead I rearchitected the entire database so the flags now go in three separate tables - create_list, modify_list, and delete_list. Each of those are nearly identical - they just hold the ids of nodes with the specified flag.

Long story short is the 100k node case impoved by 20% in initialization (making it almost on par with make!), the monitor starts in half the time, and the do-nothing update time drops to zero. So, it's super bad-ass now.

Also added and fixed t6013.