'tup g' now displays the graph the updater will see.
Before it would just display an empty graph, which is useless. It is nice to be able to see what nodes will be updated, though, even if it is a little confusing since it stuffs the create graph in with the modify/delete graph. Maybe it should spit out two graphs, somehow.