Added a bunch of tests for moving directories back and forth.
For example, moving directory a to b, then moving b back to a without doing an update in between. There was a bug in the monitor where it would just clear the old flags (since they were marked delete), and then it wouldn't update properly. That is now fixed. Also put in a few other tests, but they all seem to work.