Basically move the fchdir inside the fork in update().
The issue is when I move to multiple threads, the fchdir changes the directory for all the threads, but I only want to change the dir for the command. This unfortunately also impacted the server since it used getcwd() to figure out where the process was running. Now I added a tup_db_get_path() function to get the actual pathname from a tupid. I had to fix the canonicalization code to work with that as well. Anyway, getting pretty close to multiple threads.