Added t6015 - output files weren't removed from database.
So I ran into this trying to build libiberty. I had a bunch of commands ready to run, but didn't run them because of an early compile error. Then I started removing C files I didn't actually need, and found out later that their corresponding objects were still in the database. Everything seems to work if I check for flags & TUP_FLAGS_DELETE instead of flags == TUP_FLAGS_DELETE (since the output file was originally marked modify, when it gets removed it is modify & delete). I don't think it breaks the case I originally made that for, which was if a command is changed (so cmd A is deleted and cmd B is created, both point to the same output file, so it would be marked modify and delete). Anyway, hopefully this will work.