Added t6016 - check for removal of required input dependencies.
This can happen if you compile with an auto-generated header, and then try to remove that dependency from the Tupfile before updating. The danger would be that things could appear to be working, but someone building from a fresh checkout may have issues. Not sure if there's a better way to do it, or if it should be left out entirely. Seems kinda sensible, though.