Added t4015 - check to make sure %f args are the same order as listed.
So apparently I was putting them in backwards, so if you said the inputs are 'foo.o libbar.a', then the command would come out 'gcc libbar.a foo.o', which doesn't work. Really I just had to use list_add_tail instead of list_add. Of course that broke a bunch of tests where I was explicitly checking for a command. Whee.