I prefer having TUP_TOP always be able to have a '/' appended.
Now it can be something like "../..", or "." if you're already at the top. That way you can always do $(TUP_TOP)/foo and it will work. The main reason I changed it to this way is so you can use it like -I$(TUP_TOP) and always get a path. Before if you were at the top, it would be empty, and so the flag wouldn't work properly.