Added t6017 - don't use files that will be deleted.
So it didn't work out quite so well by combining the modify/delete flags things into a single graph. Instead I now read all of the delete stuff first, and then do modify second. The issue was gcj would read in a class file that was about to be deleted, so the dependency between java files was lost.