Made tup stop print something useful.
Now you can see if it actually tried to stop the monitor process, or if it wasn't running, or if the pid is still set but doesn't exist. The only real condition I don't check for is if I try to shut it down and for some reason the monitor is stuck. I guess that would be a bug though.