Reworked printf output.
Removed some of the spurious printfs, and moved all of the display into show_progress(). This way it works when running in parallel (since show_progress() is only called by the main thread). I didn't feel like it would be worthwhile to introduce extra lock contention just for printing, when there is already a database lock and a status lock. Also I shortened the progress bar from 40 to 11 because it's kinda silly to take up half the screen for hash marks, plus extra space for the node count, percentage, and tupid (maybe I should remove some of that, or something). Seems a little better for now.