Added t4016 (disabled) - test for chdir in a program.
So this is currently disabled because I haven't found an actual build program that does a chdir before writing its output. I'm not sure such a program would work with tup anyway, since I expect all output files to go in the directory the program was executing in (otherwise the DAG would be hosed). But if there's something that works like the shell script in this test case, then it would make sense to implement it.