Added ^ flags.
Now commands can start with text like: ^abc blah blah^ gcc -c etc... This will cause the string "blah blah" to be printed instead of the entire command string (which could be long). Basically it's like when you would do "@echo CC foo; gcc -c foo" in make, only it doesn't actually run a separate echo process (it's just a printf). Also I reserve the space where the "abc" part is for flags, in case I want to communicate something from the command to the updater. I don't know what that would be. Maybe I'll decide it's stupid and take it out. For now there just needs to be a space after the first ^. There can be any amount of whitespace after the second ^.