Reworked the message passing in the updater.
Now I just use a socketpair and write a node pointer for any worker to pick up, and then the worker writes back the node with a return code. This let me remove: 1) the silly active/inactive worker lists, 2) the node pipe for each node, 3) the status_mutex, and 4) the select on the read pipe. Hooray.