Started implementing t7004 - work properly when monitor is absent.
The goal is to be able to make some random modifications to the file system while (oops!) the monitor isn't running. Then you should be able to start the monitor and have everything work as expected (ie: without having to re-do your modifications or something stupid). I think adding new files already worked out of the box because we'd create a new node for that, which would have proper flags set and such.

I just added the ability to detect deletions. This works by copying all file/dir node ids into a separate in-memory database, and then as we walk the directories we unflag existing dirs/files. At the end, we'll have a list of all files that have been deleted while we were away. This nodes are deleted with the same treatment as tup_file_mod(), which is why I moved most of that function for deleted nodes out to tup_file_del().

The downside is it takes a bit longer to start the monitor - one of my tests with 2882 files (230 directories) went from .11s to .16s (hot caches). Assuming n*logn scaling, according to my calculations that should result in about an 8s start time with 100k nodes (assuming it could fit in a cache, which is probably an unreasonable assumption). If worst comes to worst I can probably add an option to enable/disable monitor re-start file checking.