Fixed the rest of t7004, where a file is modified.
Now in theory the whole shebang should work even if you disable the monitor for a while and make some edits. Of course when you want to run an update you'll have to wait n*logn time for the monitor to start before you can get your logn updates, but that's on you man. Anyway, I just added a config variable (64bit so it's y2038 safe or whatever) to keep track of the last time the monitor checked a file. This value also gets set when the updater finishes (the monitor turns back on). So, it seems to work pretty well. I'm sure there's some obscurebington-house bug with it somewhere. Aside, that is, from my year 2020 test case timebomb. Muwahahaha!