Database version 2: Added sticky links.
This is mainly because of t6018 - it seemed kinda stupid to get yelled at that something wasn't being specified in the Tupfile when in fact the dependency is right there. However because the command had already executed and we figured out the dependency wasn't necessary, the link was removed. So next time when the link is required, tup raises hell.

The solution is that I have added a new 'style' entry to all links. They are currently either 'normal' or 'sticky'. I chose style because type is already taken by the nodes, and I didn't feel like specifying the difference in every query. Anyway, here's the deal. Currently sticky links are only used by inputs to commands, since outputs are always required anyway. Sticky links are created and destroyed by the parser, and they reflect the actual dependencies typed in the Tupfile (including order-only prerequisites). The normal links are created and destroyed by the updater when the command is executed (like before). So sticky == Tupfile, normal == what was actually used.

They are also treated a little differently when building a DAG. The normal links are used as before - if a -> b, then when a is loaded, b must be loaded, and all of b's dependencies, etc. However, if the link is sticky, then when a is loaded, b is loaded and marked as not expanded. It's dependencies aren't loaded unless b is needed in the future. Essentially it's just left as a dead-end in the DAG, and won't be executed/used (unless something else requires b to be expanded fully). So normal links are used as before, and sticky links are used for ordering *only* if both nodes are loaded anyway.

The reason this works that way is so you could have a rule that says something like: all 'gcc -c' commands have auto-generated headers as order-only inputs. If there are any auto-generated headers that need to be updated, that will happen before any of these C files are compiled. If later one of those headers changes, only those C files that included it will be re-compiled.

Really it behaves the same way as before, only now it doesn't yell at you if you had a file that didn't include the auto-generated header, and now you do. Unfortunately I couldn't think of a way to do it without the database change :(.

End communication.