Added t6020 - check for circular Tupfile dependencies.
Since we have to parse all of a Tupfile's dependent directories before we know the output of the current Tupfile, it doesn't make sense to have directory A depend on directory B and B depend on A. So if you try to do that you'll get yelled at for having a circular dependency.

The logic is a bit weird, since it kinda relies on both the "parsing" and "already_used" flags in the node structure. Basically parsing gets set when the updater tries to parse the first Tupfile, and then "already_used" is set on the second Tupfile. When the second Tupfile is parsed and tries to parse the first again, the first node's already_used flag is 0, so parse() gets called. It then checks and sees that parsing==1, so it complains. I don't see an easy way to do it with one flag that doesn't break the ability to have one Tupfile depend on the same directory twice in two different rules, or have it silently let a circular dependency seem to work.