Changed derived files to a TUP_NODE_DERIVED type.
This type is basically TUP_NODE_FILE, but it is a file that we can safely delete. Also, it is the only type that can be the output to a command. The idea here is I wanted to protect source files from being deleted by tup when the user messed something up in their Tupfile. So now instead of deleting your C files or something, it will yell at you and say you can't write to that file. (Of course it has no knowledge of .c vs .o - it only knows if a file was put into the DAG because it was in the filesystem, or if we stuck it there because a command will write to it).

The only downside is if you removed your tup database for some reason but left a bunch of old generated files there. You'll either have to remove them by hand, or re-checkout the project.

The changes were a bit more pervasive than I had hoped - in particular, the monitor had to be changed fairly significantly in terms of renaming directories. Before I would just delete the entire old directory tree and then create a new one. Now I actually rename the directory in the DAG. This way I don't lose the knowledge of which files are base source files and which ones are derived source files. Plus it's probably a bit more efficient this way.