No longer ignore *.tup* in ldpreload.
This was used in days of yore when the individual command processes would access the database themselves (as if!). Now they don't do that anymore. I took out the check because things randomly break if you try to make a command that writes to a file called "foo.tup" or something. Not that you should do that anyway, because *.tup should only be Tupfiles that are included, and that's illegal as per t2024. However, if you're trying to write a test-case to verify such illegality, it makes things very confusing.