Added target-specific variables (t2025).
Also added t2026 and t2027 to make sure I didn't break stuff. Some other tests were modified that relied on the ability to specify %f or %o in a $-variable. However, I feel that that is much less userful that target-specific variables (which are $-variables that have a %-variable flag inside them). This mostly involved re-working the parser so I only eval pieces of the Tupfile when I need to, rather than the whole line before it gets parsed. It also means you can't put any Tupfile logic inside a variable (like: foo='include bar', then a line with $(foo) and have it include another file). But, I think that's just another benefit since it makes the Tupfiles more clear.

I also had to change the way rules were executed - now they are done as soon as they are parsed, rather than all being parsed and stuck into a list for later execution (I'm not sure why I did it that way - maybe so the whole file would be parsed before it did anything?). Anyway, this was changed so I didn't have to worry about post-setting of variables affecting a rule (like the silly deferred variables in make).