Added t3010 - make sure var/sed can be re-parsed.
Seems the logic in tup_db_is_root_node() was incorrect - instead of checking for TUP_NODE_FILE and claiming that that is a root node, I should check for TUP_NODE_GENERATED and claim that it is *not* a root node. This way TUP_NODE_VAR and such are also known to be root nodes.