Database version 3: unify normal/sticky links.
Now there is only one entry, and the style is a bitmask where 1 is normal and 2 is sticky. I ended up looking into stuff because the performance of parsing a Tupfile with a lot of links (specifically, a test linux/fs/ext3/Tupfile) was kinda crappy. It took about 420ms to parse the Tupfile, which seemed kinda excessive. Most of the time was spent in tup_db_get_src_links(). I think the reason it was slow is because the 'select...from_id not in (select id from delete_list where id=from_id)' was being executed before the rest of the query, so it was going through every link. Some bizarre permutations made it run at 170ms, though I admit I don't understand why. Anyway, I didn't figure that out until after I had already re-written everything to use a new tup_db_yell_links call so all the searching for bad links is done in sqlite. I like this approach better because it got rid of my nested list_for_each_entrys() (O(n^2)! - the exclamation point shows surprise, not factorial). Also, it brings the parsing time of fs/ext3 down to 140ms, with no readily apparent single point of slowness.

The downside to the yell query (so called because it yells at you if it finds any rows) is that I could only think of a way to write it by making the style a bitmask, which unfortunately requires a new database version. Fortunately noone uses tup yet, so I think we're safe...this time...