Database version 4: Added support for symlinks
Previously, symlinks didn't really work correctly in that commands weren't dependent on both the symlink node and the node it points to. Now they are, as well as any symlinks for directories that were used to read in files.

Unfortunately I didn't break this commit up into smaller chunks (I tried a few times, but it's pretty annoying). I started by just trying to add some symlinks, then realized later I needed a few infrastructure-y type changes, so those are in there too. Specifically, the best way to handle symlinks seemed to be by adding a column to the node table to reference a linked-to tupid. I tried to put symlinks in the link table (with a special style), but that made it difficult to get a list of symlinks (basically doubling the number of queries). I also tried to put the symlink-finding in canonicalize(), but that was stupid because it kept calling stat() on every path when it was canonicalizing to see if it needed to do readlink(). Instead I opted to change out find_dir_tupid_dt() works by letting it to a while loop over the new sym column in the node table, which it already has access to because it's pulling in the node anyway.

While doing this, I realized I can probably get rid of canonicalize in favor of a smarter find_dir_tupid_dt, but I haven't finished that yet. I wanted to get these sweeping reforms into a commit before it gets even crazier :).

Currently t6017 is broken because it writes to a full path, but that should be fixed once I finish the canonicalization->find_dir_tupid_dt changes.