Added t4017 - another canonicalization issue.
Before I was canonicalizing everything in server.c, but now I'm not. So in order to properly compare if files are equal I had to split the files up into path elements and compare them (for example, when temporary files are unlinked and need to be ignored). I still need to refactor write_files() a bit because it will use find_dir_tupid_dt() and such which will re-break the path up into path elements, which is silly because I already have them.