Change some of the benchmarks so they don't create files.
In some cases I just want to test adding nodes to the database, not files in the filesystem. Since the touch() command was changed to actually create the file, this hosed the benchmarks that were just adding nodes (in particular when I run them with 100k or 1M nodes). Now there's a separate 'tup node' command that can just add a node to the database.