Added t7014 - symlink shouldn't cause updates when monitor starts.
The update_symlink_file() is called by the monitor on startup when it goes through the directories. This function was always calling tup_db_add_modify_list(), so just starting the monitor would cause things to get updated, which is silly. Now update_symlink_file() only causes it to be put in the modify list if the symlink actually changed (or didn't exist to begin with). This also means in touch() I had to call tup_file_mod() on the symlink in case the symlink already existed and points to the same file.