Initial work for ghost nodes.
These are nodes that a program attempts to read from, but don't correspond to actual files. However, if the file is later created, the commands that attempted to read from them should be re-executed. A ghost node is created if a program either opens a file or stats a file and fails with ENOENT. Note that stat() doesn't count as reading a file if it exists, since file metadata isn't considered a part of the build state (ie: a program shouldn't use file metadata for anything). However, a program may do something like 'if stat(file), then open(file)' so it makes sense to create a ghost node if the stat fails with ENOENT.

Currently things don't work properly if the ghost node is in a ghost directory (not sure how to handle that yet). Also, ghost nodes aren't deleted yet, unless they are converted to regular file nodes and then removed. Still need to figure out the best approach for that.