Added t5023 - slightly realistic ghost lib test.
I remember when I added this error case in todo that I didn't think tup would ever be able to handle it properly, and it would be a caveat for all eternity. Well, after harnessing the awesome power of the Secret of the Ghost Valley, I can mark this one off the list. W00t.