Added t5024 - symlink to non-existent file.
Had to make sym_follow public so it could be called when processing ghost nodes in write_files() (since a read on a symlink pointing nowhere will fail to open, it will show up as a ghost node). The way it was structured, I also needed to follow the symlink of the file itself (normally this is done in get_dbn_*(), but that doesn't work here...possibly a candidate for refactoring).

Also had to make touch() only create a file if the lstat fails - otherwise if you touch a symlink pointing to a non-existent file, it ends up creating the file, which I didn't want for this test.