Database version 5: Adjust link_index to (from_id,to_id)
Previously, link_index was just on from_id. Since inserting a link requires a query in order to get the style (and for uniqueness), it was slow to only use an index on from_id. Specifically, after I made mlink() run in a transaction and fixed up the b11-link.sh benchmark to not run so many processes, the benchmark on sizes [100, 1000, 10000] was at [0.09s, 1.43s, 130.98s]. After the link_index change, it is at [0.09s, 0.25s, 1.94s]. A 100k link test now runs at 20.12s (I didn't bother to wait for 100k links before...)