Use a variable dictionary instead of a socket.
Now whenever variables are changed, the .tup/vardict file is updated. It's in alphabetical order and has an index, so it's easy to do a binary search on the file to find a particular variable. This way a tup-itized gcc can read from this file when it needs a variable, and then just notify the wrapper that it used a particular variable. The original method involved sending a message up, and the wrapper process would lock the database, read the variable's value, unlock the database, and send the result back. This seemed silly, as it would result in lots of lock contention for the database when running in parallel, and also involved a number of context switches.

The new method just passes in a file descriptor to an already-opened .tup/vardict file that all subprocesses can share read-only access to. Any variables accessed are passed in via the socket in the same manner that accessed files are. The variable tupids are linked at the same time as actual file accesses.

The reason I initially tried to do this is because my test on linux/fs/ext3/*.c was running at 18.5s with tupgcc, whereas tup with autoconf.h was at 11.8s. Just this change ended up bringing the time down to 17.5s, which wasn't as good as I had hoped. However, I found that the extra 6s was really because of my crappy link index, which I just fixed on the master branch. So with the link fix and the variable dictionary here, I'm now at 10.1s to compile all the C files and link built-in.o (21 commands in all). The absolute minimum using an oracle script and autoconf.h is 8.684s, whereas an oracle script using linux's fixdep mechanism is at 8.944s. So I'm only at about 13% overhead over kbuild, if you assume you can use an oracle script. Running make in the linux tree for me takes 15.3s to do nothing with a hot cache, so really tup can build the ext3 subtree 33%-50% faster than kbuild can do nothing, depending on how you divide ;)