Database version 6: Add index on node.sym
Also added tests t2032, t2033, and t5025. These are all more complex ghost node cases, with respect to ghost nodes in ghost directories, ghost directories getting created as files, and symlink destinations that get removed (which is what required the database update).

Currently this breaks a few tests - specifically, ones that delete directories and then check that the directory node is removed from the database. This doesn't work at the moment since the directory nodes hang around as ghosts. I'm still trying to figure out a good way to do that, but since it may be tricky I wanted to get these changes in first so I have room to play.