Database version 7: Add ghost_list.
Added a bunch of tests to make sure ghosts are reclaimed when they are no longer referenced, and stick around if they are still referenced. A ghost should be alive if another node points to it in its dir or sym field, or if the ghost is used in the from_id (to point to a command). Now when a node is deleted, it will add its dir and sym references to the ghost_list. When tup is about to shutdown, it will go through the ghost_list and purge any ghosts that no longer have any references.

The 'tup delete' command was also slightly adjusted so you can actually delete a symlink. Before it was following the symlink path too far and trying to delete the node it points to.