Added t2040, t5029, t5030 - more symlinking/ghosts.
Made sure that a rule can generate a symlink to a ghost node. Also I found out that ghost dirs with ghosts in them may not get deleted in all cases, so I made t2040 to check for that. Just before deleting ghosts, I have it insert all ghost node's dirs (if they are ghosts) into the list.