Added t5032 - accidentally write out a bad symlink.
This was a bit problematic because the symlink would hang around and then cause the command that was re-executed to fail (since it would be writing through the symlink). I fixed that by moving delete_file() from the updater to a globally accessible spot, and then call that to remove the bad output file. Some of the internals of write_files() were re-worked, but it's still pretty messy. It should properly identify and remove all bad outputs, whether they be actual files or symlinks.