Added t5033 - make sure symlink cycles work-ish.
Since a symlink cycle returns ELOOP when trying to open the file, we never see it in write_files(), so that's good. However, we still have a cycle in the DAG in the database. I had to reset the sym entry in the node when converting a node to a ghost in order to break the cycle so nodes could be removed.