Database version 8: Remove ghost_list.
Turns out putting ghost_list into the database was silly and unnecessary. It would probably just add to database fragmentation. Since it really is only built up during a transaction and cleared out before it ends, I ended up making it a temporary table, and defining SQLITE_TEMP_STORE to 2 so that this happens in memory. By my calculations, this is faster than using SQLITE_TEMP_STORE=1 (the default), and also faster than attaching a new database in memory and creating a table. As such, it may be prudent to rework tmpdb to behave similarly. Also added t7016, since I wasn't properly removing ghosts in the monitor (which prompted me to put that someplace global, when transactions are committed).