Variable deletion now handled with a temporary list.
Before I would mark all the variables in the delete list, then have to do a weird query to mark dependent dirs, and all the variable queries had to make sure that the var wasn't in the delete list. Instead I now copy all the var ids into a temporary list, and then all the deletion operations work on that list in the end (which is nice, because we know in advance that they are all variables). In addition, this means that a variable will never be marked delete, so all those queries are simplified.