Added t3012 and t3013 - ghost variables.
Now if a Tupfile tries to read from a variable that doesn't exist, instead of an error it will become a ghost variable, which is treated as empty. The Tupfile will still be re-parsed later if the variable is created, and the ghost variable goes away if the Tupfile no longer reads from it.

Also fixed up the silly query in tup_db_delete_dependent_dir_links(). It's kinda funny how stupid it was. Unless it did something really important and I'm just not seeing it.