Changed t6017 so it's not dependent on gcj.
I got quite tired of waiting around for the 5 seconds it takes it to compile a 3-line java file. This mimics the javac behavior by doing an ls and grepping for the file it wants (as opposed to just stating the file and checking the return value). I verified that the test failed in commit aadce2e6^ and worked in aadce2e6.