Change dollar-sign behavior to not require backslash.
It was annoying because any $ symbol that was supposed to go to the shell had to be escaped. Now it only requires escaping if you want to send $(foo) to the shell. In that case you need to type \$(foo). But things like 'sed s/foo$/bar/' are much easier to copy/paste from a real shell.