Added t5034 - changed hidden file behavior
Now you can write to a hidden file in a command, but you get a warning (and a summary of the warning count is printed at the end of the update). Not sure if that's the most practical way to handle it, but I figure it might help sometimes to write stuff out to a file without letting tup know about it during a debugging session.

Also, hidden files are now distinguished from out-of-tree files (like /usr/bin/gcc). Trying to read from hidden files is an error, but out-of-tree files go untracked and unreported.