Don't use duplicate commands anymore.
Now I just save all the files written to/read from in a temporary list of tupids, and then use a few global queries to check for errors and update the existing command node. This way I don't have to churn the database so much by deleting an old command and hundreds of links, only to replace it with the same command and same hundreds of links. The test cases themselves only went from 17.5s to 17.0s, which isn't that great. But touching linux/fs/ext3/balloc.c went from 3.718s to 2.486s - quite a jump! Especially considering that the oracle solution is 1.759s - this puts tup at 41% overhead. The majority of that is actually the commit writing to disk. If you disable syncing it goes down to 1.986s, which puts me at 12.9% overhead :)