Kill all subprocesses on a second signal.
Now the children (and tup itself) are SIGKILLed if you ctrl-c tup twice. The first time it will wait for the current children to finish (like before). Now the second time it will make sure they are killed.

Also, after the fork(), the child will close its file descriptors that it has on the lock files. At least in my "sleep 60" test, even kill -9'ing the sleep process still seemed to hold on to the file descriptors until the sleep period ended (I guess the kernel didn't finish the process until the delay was over). The annoying thing is then a new tup upd won't start until that ends. Closing the lock files fixes that. Of course it can't call the existing tup_lock_exit() since that explicitly calls flock(LOCK_UN), which would unlock those and return control to the monitor, which would be bad.

The only possibly annoying thing about this implementation is that tup won't save any of its progress, since it sends SIGKILL to itself as well.