Added t7021 - make sure symlink in subdir works for the monitor.
This is mostly a refactoring in preparation for adding a 'tup scan' command. Now I always have an open file descriptor for the top-level directory which can be retrieved from tup_top_fd(). Also I was able to remove the full path information from dircache, which means it just does a wd -> dc mapping. That may mean dircache is an extra unnecessary layer now. Anyway, that cleaned up the watch_path() code quite a bit, which is nice. Also it makes the monitor start up ~15% faster, which is nice.

After the initial refactoring, all test cases passed, but the monitor failed to start in my gittup/ test directory because of a symlink file in a subdirectory. I added t7021 to make sure that doesn't break again (since it was working before, obviously).