Database version 9: Add mtime.
In order to support a full filesystem scan, I added a last-changed mtime to the database. Now when the startup scan happens in monitor, or is explicitly executed via 'tup scan', the timestamps of the files are checked against the stored mtimes. Those that differ are assumed to have been modified. This was done because it is near impossible to correctly implement the scanner with a single global timestamp (like the monitor had).

Note that though it uses mtime, it is vastly different than how make would do it, because it is not comparing timestamps of dependent files. Rather, it compares the timestamp to the last known timestamp for that file.

The 'tup scan' implementation is O(n), but is just there as a basic file modification detector for platforms where the monitor isn't supported, or perhaps isn't possible. (Though in order to truly benefit from tup, a monitor or filesystem support should probably be added for those platforms). Tup still beats the snot out of make even with this, though.