Added t2047 - multiple rules output into the same bin.
Also while checking this I realized it was possible to have duplicate links on input files. The issue was that I returned 0 in tup_db_link_style to indicate both database success (non-existent link), and a link that exists with style of 0. This causes a problem if the link is only a sticky link, so when I unsticky it in order to update a command, the style becomes 0 temporarily, and then I add a new link instead of updating the existing one.

In order to check for this sort of thing cropping up again (especially since I may re-do how the parser handles all of this), I added a check_dup_links command that can find and print duplicate links (returning an error if there are any). This gets checked now in a few places in the test scripts, as well as after any test script finishes.