Added t6023 - fix annoying ldpreload issue.
So apparently I can't rely on the constructor getting called before my functions are actually linked in and wrap the libc calls. In particular, this seems to happen if another shared library linked directly into the program uses libc calls (like fopen) in one of its constructors. Instead now I just check each time if the function pointer is set or not (and similarly for the socket when sending info).