Added t2053 - shell globbing test.
I had to change bins to use {} instead of [] because the latter is used by sqlite's glob function, and it might be useful to use that globbing feature at some point. Plus it would be a pain to change the bin syntax later, or have to escape one or the other in order to use both in an input list.