Removed TUP_TOP and include_root in favor of TUP_CWD.
The issue is any Tupfile using TUP_TOP or include_root is inherently not compositable. So if someone wants to incorporate your project as a sub-project, all of a sudden TUP_TOP is pointing to the wrong place, and no include_root Tupfiles would be found. Instead I have TUP_CWD which can be set to variables (like LINUX_ROOT) or something. That currently isn't very useful until I put in an include_rules directive, but that should be coming next.